рефакторинг редьюсеров
