+++ project/src/components/controls/controls.tsx
@@ -3,7 +3,8 @@ import { useSelector, useDispatch } from 'react-redux';
 import { Link, useHistory, useParams } from 'react-router-dom';
 import { AuthorizationStatus, FavoriteStatus } from '../../const';
 import { fetchFavoriteStatusAction } from '../../store/api-actions';
-import { getFilm, getPromo } from '../../store/films/selectors';
+import { getFilm } from '../../store/film/selectors';
+import { getPromo } from '../../store/promo/selectors';
 import { getAuthorizationStatus } from '../../store/user/selectors';
 
 type ControlsProps = {
