+++ project/src/pages/add-review-screen/add-review-screen.tsx
@@ -3,7 +3,7 @@ import Breadcrumbs from '../../components/breadcrumbs/breadcrumbs';
 import Header from '../../components/header/header';
 import CommentForm from '../../components/comment-form/comment-form';
 import ErrorScreen from '../error-screen/error-screen';
-import { getFilm } from '../../store/films/selectors';
+import { getFilm } from '../../store/film/selectors';
 
 function AddReviewScreen(): JSX.Element {
   const film = useSelector(getFilm);
