+++ project/src/pages/main-screen/main-screen.tsx
@@ -17,11 +17,11 @@ import {
 } from '../../store/films/selectors';
 import {
   getErrorFilmsStatus,
-  getLoadingFilmsStatus,
-  getLoadingPromoStatus
+  getLoadingFilmsStatus
 } from '../../store/films/selectors';
 import { fetchPromoAction } from '../../store/api-actions';
 import { getAuthorizationStatus } from '../../store/user/selectors';
+import { getLoadingPromoStatus } from '../../store/promo/selectors';
 
 function MainScreen(): JSX.Element {
   const genres = useSelector(getGenres);
