+++ project/src/store/app/selectors.ts
@@ -2,7 +2,7 @@ import { NameSpace } from '../root-reducer';
 import { State } from '../../types/state';
 
 export const getActiveGenre = (state: State): string =>
-  state[NameSpace.UpState].activeGenre;
+  state[NameSpace.App].activeGenre;
 
 export const getLimit = (state: State): number =>
-  state[NameSpace.UpState].limit;
+  state[NameSpace.App].limit;
