+++ project/src/store/root-reducer.ts
@@ -1,20 +1,32 @@
 import { combineReducers } from 'redux';
 import { films } from './films/films';
-import { upState } from './up-state/up-state';
+import { app } from './app/app';
 import { reviews } from './reviews/reviews';
 import { user } from './user/user';
+import { favorites } from './favorites/favorites';
+import { film } from './film/film';
+import { promo } from './promo/promo';
+import { similarFilms } from './similar-films/similar-films';
 
 export enum NameSpace {
   Films = 'FILMS',
+  Film = 'FILM',
+  Promo = 'PROMO',
+  SimilarFilms = 'SIMILARFILMS',
+  Favorites = 'FAVORITES',
   Reviews = 'REVIEWS',
-  UpState = 'UP_STATE',
+  App = 'APP',
   User = 'USER',
 }
 
 export const rootReducer = combineReducers({
   [NameSpace.Films]: films,
+  [NameSpace.Film]: film,
+  [NameSpace.Promo]: promo,
+  [NameSpace.SimilarFilms]: similarFilms,
+  [NameSpace.Favorites]: favorites,
   [NameSpace.Reviews]: reviews,
-  [NameSpace.UpState]: upState,
+  [NameSpace.App]: app,
   [NameSpace.User]: user,
 });
